Add Docker image for RHEL 10 (#59)
This change adds support for RHEL 10. At the moment no gcc-toolsets exist,
and RHEL 10 comes bundled with GCC 14.

Assuming gcc-toolsets will be added in the near future, this change
therefore checks whether the image built is RHEL 9 (in which case it
will install the requested GCC version using the gcc-toolset), or
otherwise will install the default compiler (until some day we can
revert this piece of conditional logic to also use the toolset).