Update READMEs (#69)
This PR updates the READMEs as follows:
* Logging into GHCR is only needed when pushing, so those instructions
are moved to the last section.
* GCC base image ARG was missing ghcr.io, which has been fixed.
* To simplify copy-pasting of commands, the `CONTAINER_REGISTRY` var has
been removed and ghcr.io has been added explicitly. Otherwise it's
needed to copy-paste-run commands scattered across the readme.
* Extra note for macOS is included to add `--platform linux/amd64` in
order to create a Docker image that can be run by CI on Linux.
* Updated push instruction to include `sha-${USER}` to the image tag, so
it can be referenced in the strategy matrix.
* General other cleanups.