Add tools image for .deb and .rpm signing (#76)
Adding the ability to sign the .deb and .rpm packages to the CI builder
images is problematic for RHEL. Since it actually isn't necessary to use
the same image to sign a package created by a particular image, this PR
adds the signing packages to a new "tools" image.

As an added bonus, it cleans up the documentation image by adding the
`--no-install-recommends` flag.