Add ccache (#85)
This change adds `ccache` to the Debian, Ubuntu, and RHEL images, so our
CI pipelines can take advantage of caching (once configured to do so).

Note that due to our use of the UBI images of RHEL, the `ccache` package
cannot be installed using the default repositories available through the
package manager. I first tried installing it by pulling it from from
GitHub, however that required installing an appropriate versioning for
each RHEL distro as otherwise GLIBC incompatibilities would surface.
Instead, I added the RHEL version-specific Fedora EPEL repository, which
is a project that provides RHEL-compatible packages, as that will
install a compatible `ccache` version.

To get a better view of which version is installed of which package,
including `ccache`, they are now printed together at the end of each
base image. This also acts as a sanity-check that the packages are
actually installed.